+++ R/.old/_refactor/refactored_but_obsolete/PackageRepository.S3.r
@@ -1,0 +1,38 @@
+#' @title 
+#' Constructor for package repository class
+#'
+#' @description
+#' Creates an instance of class \code{PackageRepository.S3}.
+#'
+#' @template intended-use
+#'
+#' @param .x \code{\link{ANY}}. An object of an arbitrary class whose class
+#'    attribute should be updated so that it becomes an instance of class
+#'    \code{PackageRepository.S3}. Mainly intended for rapid prototyping 
+#'    purposes. As the development stage matures, the distinctive fields should
+#'    be used.
+#' @field path \code{\link{character}}. Installation url.
+#' @field type \code{\link{character}}. Application type.
+#' @return Instance of class \code{PackageRepository.S3}.
+#' @example inst/examples/PackageRepository.S3.r
+#' @template author
+#' @template references
+#' @keywords class, constructor
+#' @export
+PackageRepository.S3 <- function(
+  .x,
+  root = NA_character_,
+  sublevel = list()
+) {
+  cname <- "PackageRepository.S3"
+  if (!missing(.x)) {
+    class(.x) <- c(cname, class(.x))
+    out <- .x
+  } else {
+    out <- structure(
+      list(root = root, sublevel = sublevel),
+      class = c(cname, "list")
+    )
+  }
+  return(out)
+}