+++ inst/examples/old/PackageRepositorySubGeneric.S3.r
@@ -1,0 +1,4 @@
+PackageRepositorySubGeneric.S3()
+PackageRepositorySubGeneric.S3(path = "a/b/c")
+
+PackageRepositorySubGeneric.S3(.x = c(TRUE, FALSE))