+++ inst/examples/old/PackageRepositorySubSource.S3.r
@@ -1,0 +1,4 @@
+PackageRepositorySubSource.S3()
+PackageRepositorySubSource.S3(path = "a/b/c")
+
+PackageRepositorySubSource.S3(.x = c(TRUE, FALSE))