+++ inst/examples/old/PackageRepositorySubWindows.S3.r
@@ -1,0 +1,4 @@
+PackageRepositorySubWindows.S3()
+PackageRepositorySubWindows.S3(path = "a/b/c")
+
+PackageRepositorySubWindows.S3(.x = c(TRUE, FALSE))