+++ rapp/build.r
@@ -1,5 +1,5 @@
 ## Working directory //
-path_wd <- "q:/home/wsp/rapp2/repositr"
+path_wd <- "q:/home/wsp/rapp2/repor"
 if (getwd() != path_wd) {
   setwd(path_wd)    
 }